requirements.txt
PyQt6
requests
pyodbc
pandas
opencv-python
opencv-python-headless
face_recognition
matplotlib
seaborn
PyMuPDF
PyPDF2
pdfplumber
tabula-py
camelot-py
pytesseract
pdf2image
pillow